+++ src/routes/api/leaderboard/+server.ts
@@ -43,7 +43,7 @@ export const GET: RequestHandler = async ({ url }) => {
 
 			return {
 				atoms: parseFloat(entry.atoms), // Use parseFloat instead of parseInt
-				equippedSkin: entry.equipped_skin ?? null,
+				equippedBanner: entry.equipped_banner ?? null,
 				level: entry.level,
 				is_online: isTrulyOnline,
 				picture: entry.picture || '',